Add ability to run tpm2_makecredential without a TPM
Used some functions from tpm2_import for off-TPM functionality
 - Move needed, overlap code into tpm2_identity_util
Add new global -X/--openssl-backend option for any tools that can operate w/o TPM

Need to use full-sized TPM2B_MAX_BUFFER instead of TPM2B_DIGEST
Make sure 'tool_opts' exists in tpm2_options before using it
Add test for --openssl-backend mode in makecredential integration suite

Fixes for SAPI-less makecredential PR

Clean up code, remove commented out code
Remove debug-ish tool output
Changed --no-tpm option to --openssl-backend
