+++ lib/tpm2_options.h
@@ -45,6 +45,7 @@ union tpm2_option_flags {
         UINT8 verbose : 1;
         UINT8 quiet   : 1;
         UINT8 enable_errata  : 1;
+        UINT8 no_tpm  : 1;
     };
     UINT8 all;
 };
